Lobby & Auth

+++ src/user/user.controller.ts
@@ -1,34 +1,39 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Patch,
+  Param,
+  Delete,
+  UseGuards,
+  Request,
+} from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
+import { AuthGuard } from 'src/auth/auth.guard';
 
 @Controller('user')
+@UseGuards(AuthGuard)
 export class UserController {
   constructor(private readonly userService: UserService) {}
 
-  @Post()
-  create(@Body() createUserDto: CreateUserDto) {
-    return this.userService.create(createUserDto);
+  @Get('')
+  async findOne(@Request() req) {
+    if (!req.user.data.sub) throw new Error('No UID provided');
+    return await this.userService.findOne(req.user.data.sub);
   }
 
-  @Get()
-  findAll() {
-    return this.userService.findAll();
+  @Patch('')
+  async update(@Request() req, @Body() updateUserDto: UpdateUserDto) {
+    if (!req.user.data.sub) throw new Error('No UID provided');
+    return await this.userService.update(req.user.data.sub, updateUserDto);
   }
 
-  @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.userService.findOne(+id);
-  }
-
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
-    return this.userService.update(+id, updateUserDto);
-  }
-
-  @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.userService.remove(+id);
+  @Delete('')
+  async remove(@Request() req, @Param('id') id: string) {
+    if (!req.user.data.sub) throw new Error('No UID provided');
+    return await this.userService.remove(+id);
   }
 }
